+++ android/src/main/java/com/dieam/reactnativepushnotification/modules/RNPushNotification.java
@@ -8,8 +8,8 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Bundle;
-
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.core.app.NotificationManagerCompat;
 
 import com.dieam.reactnativepushnotification.helpers.ApplicationBadgeHelper;
@@ -28,6 +28,7 @@ import com.facebook.react.bridge.WritableMap;
 
 import java.io.IOException;
 import java.security.SecureRandom;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -35,12 +36,20 @@ import android.util.Log;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.iid.InstanceIdResult;
 import com.google.firebase.messaging.FirebaseMessaging;
 
 public class RNPushNotification extends ReactContextBaseJavaModule implements ActivityEventListener {
     public static final String LOG_TAG = "RNPushNotification";// all logging should use this tag
+    public static final String KEY_TEXT_REPLY = "key_text_reply";
+
+    public interface RNIntentHandler {
+        void onNewIntent(Intent intent);
+  
+        @Nullable
+        Bundle getBundleFromIntent(Intent intent);
+    }
+  
+    public static ArrayList<RNIntentHandler> IntentHandlers = new ArrayList();
 
     private RNPushNotificationHelper mRNPushNotificationHelper;
     private final SecureRandom mRandomNumberGenerator = new SecureRandom();
@@ -61,7 +70,7 @@ public class RNPushNotification extends ReactContextBaseJavaModule implements Ac
 
     @Override
     public String getName() {
-        return "RNPushNotification";
+        return "ReactNativePushNotification";
     }
 
     @Override
@@ -81,6 +90,12 @@ public class RNPushNotification extends ReactContextBaseJavaModule implements Ac
             bundle.putBundle("data", intent.getExtras());
         }
 
+        if (bundle == null) {
+            for (RNIntentHandler handler : IntentHandlers) {
+                bundle = handler.getBundleFromIntent(intent);
+            }
+        }
+
         if(null != bundle && !bundle.getBoolean("foreground", false) && !bundle.containsKey("userInteraction")) {
           bundle.putBoolean("userInteraction", true);
         }
@@ -90,12 +105,16 @@ public class RNPushNotification extends ReactContextBaseJavaModule implements Ac
 
     @Override
     public void onNewIntent(Intent intent) {
+        for (RNIntentHandler handler : IntentHandlers) {
+            handler.onNewIntent(intent);
+        }
+        
         Bundle bundle = this.getBundleFromIntent(intent);
         if (bundle != null) {
             mJsDelivery.notifyNotification(bundle);
         }
     }
-    
+
     @ReactMethod
     public void invokeApp(ReadableMap data) {
         Bundle bundle = null;
@@ -118,17 +137,17 @@ public class RNPushNotification extends ReactContextBaseJavaModule implements Ac
     public void requestPermissions() {
       final RNPushNotificationJsDelivery fMjsDelivery = mJsDelivery;
       
-      FirebaseInstanceId.getInstance().getInstanceId()
-              .addOnCompleteListener(new OnCompleteListener<InstanceIdResult>() {
+      FirebaseMessaging.getInstance().getToken()
+              .addOnCompleteListener(new OnCompleteListener<String>() {
                   @Override
-                  public void onComplete(@NonNull Task<InstanceIdResult> task) {
+                  public void onComplete(@NonNull Task<String> task) {
                       if (!task.isSuccessful()) {
                           Log.e(LOG_TAG, "exception", task.getException());
                           return;
                       }
 
                       WritableMap params = Arguments.createMap();
-                      params.putString("deviceToken", task.getResult().getToken());
+                      params.putString("deviceToken", task.getResult());
                       fMjsDelivery.sendEvent("remoteNotificationsRegistered", params);
                   }
               });
@@ -207,11 +226,11 @@ public class RNPushNotification extends ReactContextBaseJavaModule implements Ac
 
     @ReactMethod
     /**
-     * Cancel scheduled notifications, and removes notifications from the notification centre.
+     * Cancel scheduled notification, and remove notification from the notification centre.
      *
      */
-    public void cancelLocalNotifications(ReadableMap userInfo) {
-        mRNPushNotificationHelper.cancelScheduledNotification(userInfo);
+    public void cancelLocalNotification(String notification_id) {
+        mRNPushNotificationHelper.cancelScheduledNotification(notification_id);
     }
 
     @ReactMethod
@@ -261,17 +280,8 @@ public class RNPushNotification extends ReactContextBaseJavaModule implements Ac
      * Unregister for all remote notifications received
      */
     public void abandonPermissions() {
-      new Thread(new Runnable() {
-          @Override
-          public void run() {
-              try {
-                  FirebaseInstanceId.getInstance().deleteInstanceId();
-                  Log.i(LOG_TAG, "InstanceID deleted");
-              } catch (IOException e) {
-                  Log.e(LOG_TAG, "exception", e);
-              }
-          }
-      }).start();
+      FirebaseMessaging.getInstance().deleteToken();
+      Log.i(LOG_TAG, "InstanceID deleted");
     }
 
     @ReactMethod