+++ component/index.android.js
@@ -2,7 +2,7 @@
 
 import { NativeModules, DeviceEventEmitter } from "react-native";
 
-let RNPushNotification = NativeModules.RNPushNotification;
+let RNPushNotification = NativeModules.ReactNativePushNotification;
 let _notifHandlers = new Map();
 
 var DEVICE_NOTIF_EVENT = 'remoteNotificationReceived';
@@ -32,8 +32,12 @@ NotificationsComponent.prototype.subscribeToTopic = function(topic) {
 	RNPushNotification.subscribeToTopic(topic);
 };
 
-NotificationsComponent.prototype.cancelLocalNotifications = function(details) {
-	RNPushNotification.cancelLocalNotifications(details);
+NotificationsComponent.prototype.unsubscribeFromTopic = function(topic) {
+	RNPushNotification.unsubscribeFromTopic(topic);
+};
+
+NotificationsComponent.prototype.cancelLocalNotification = function(details) {
+	RNPushNotification.cancelLocalNotification(details);
 };
 
 NotificationsComponent.prototype.clearLocalNotification = function(details, tag) {